Introduce fluid effect for fiery - melt block
Melts the block based on its melting recipes to produce a fluid block.
In most cases, this gives you more fluid than its supposed to and is only balanced as fiery is so rare.
However, the fluid effect is configurable so you can make it require more fluid  before it does its thing, or filter which blocks it works on