Replace electrum's trait with shock
Shock causes bonus damage to targets in water or rain. On armor, it causes thorns-like damage to targets in water or rain
To implement this, includes the new "water" entity variable that returns 2 for water, 1 for rain. Could have done that via composiing existing effects but this seemed cleaner
Also includes a new damage type parameter for the thorns module