Phase 1 of improving visuals for swimming in fluids
Fluids now all have camera textures, which are generated automatically from the first frame of the fluid texture.
For now while I have not settled on fog behavior, the camera textures are drawn at 75% opacity, making it similar to when you head is inside a block but not fully restricting sight. Will turn down that number on fluids that get fog.