Improve travelers gear recycling
The curiass on travelers now maps to a maille, since maille has a superset of curiass materials
The travelers shield core now maps to a repair kit; shield core from recycling was a material dupe. Required a recipe change, so I took the chance to make it a 2x2 recipe.