Improve grapple formula
My original constant was specific to the target distance of about 11 blocks. Now instead computing it as the inverse square root of the distance you wish to travel, with a constant to calibrate that.
Notably makes it stronger at close range, and weaker at far range, which means at most distances you are within a half a heart of fall damage of landing on top. Worked super well in some parkour testing in the twilight forest.