Improve ToolAttackUtil via a ToolAttackContext builder
Makes it far easier than the half a dozen overloads, plus more added with the new projectile stuff. Will make it easier to add new stuff in the future as well.
As part of this, fixed the issue where modifiers do not receive updated critical hit information from events
Also fixed a crash with tools with attributes on the projectile, as it turns out its unsafe to fetch projectiles at that time. May revisit with an altenative algorithm but leaning towards no attributes for trident consistency