Add recipes for throwing and returning
Can be added to any tool supporting chargable modifers, other than bows. Created a new tag to faciliate such requests
Throwing is crafted using a trident, but also available as as trait on javelins
Returning (renamed loyalty) is crafted using a totem of undying, though higher levels just cost a clock (and an upgrade slot).
Due to the recipe cost difference, returning is not extractable using modifier crystals, though the enchantment is a fine way to obtain it.
Also added an upgrade blacklist for book extracting, since the modifier is both ability and upgrade