Rebalance projectiles
Shurikens now have 1.5 base damage, but a 2x power multiplier. This makes power not basically useless on shurikens at the cost of basic shuriken damage.
Glass's smashing trait now has a downside: -0.75 power. Makes up for how big some of those bonuses get and encourages using it for low velocity bows
Crystalbound's power bonus is now just +0.75, as the condition is an easy one to achieve on an inexpensive material. Puts it more in line with quartz as an interesting trade off.
Overall, this makes shuriken damage land about where I want it: decent bit of effort to reach 7 damage