Reorder encyclopedia index
First, hide the introduction so we are back down to 12 sections.
For row 1, list out general tools with melee/harvest, ranged, and ammo materials
For row 2, start with armor, armor materials, and slimeskull materials.
For row 3, list all 3 modifier types.
Fluid effects ended up at the end of row 2 since that was the one row that had space leftover. Its debatably a material type so it fits there.