Switch BreakSpeedModifierHook argument from the event to a custom context
Will make it easier for thrown tools to break blocks
For compatability, keeping the old methods around as deprecated for removal. Did my best to migrate to the event automatically, but will be way more performant to update in addons
Skipped the two deprecated for removal usages of the modifier hook though, no one should be using them anyways.