Fix necrotic not applying to fishing rods
Closes #5521
Only worked on arrows before, now works on any projectile.
This is making the lower base damage of fishing rods stand out more, perhaps we should move that to a damage multiplier instead of hardcoded. Though tbh, it is pretty easy to necrotic from rods so maybe its better this way. Also considering the way it impacts rod explosions.