Implement shaped parts recipe
SImilar to shaped material recipe, but accepts a list of tool parts to search for and update the output.
Could be used for say, crafting a pickaxe or hammer in the crafting table, though we don't plan to do that in the core mod.
For our uses, will be an alt travelers gear recipe using a fake part.