Implement material block model
Used for making 3D tool parts and tools. The 3D tool part that this will be used for is kinda disappointing though.

Notably does not support modifiers, I have not yet decided on how to do those.
Also does not support the bow ammo rendering, that requires mixing model types and is also complex.