Make modifier requirements module only run on crafted modifiers by default (#5530)
I cannot think of a practical usecase for wanting it to run on traits, and a user made a good argument for why its useful to have it not run.
Implemented through a new modifier hook, validate upgrade. Works the same as the validate hook but is run on the upgrade list.