Tweak a few modifiers
Ductile now grants +0.25 power instead of +5%. Works out to be the same on a power 5 tool, better with less power.
Featherweight, lightweight, and heavy both perform additive instead of multiplicative ranged stats. This is on average better for bows with smaller stats, but weaker on large stats.
Maintained now grants +10% velocity per level instead of +5%
Also took the chance to correct descriptions for cobalt and skyslime vines in the books and tooltips